+++ src/components/form/groups/BulmaFormControlGroup.vue
@@ -1,6 +1,5 @@
 <template>
 	<div :class="classes" class="field">
-		<!--todo: only holds 1 input/control within the control class-div; make a prop for multiple controls to comply with has-addons class-->
 		<slot />
 	</div>
 </template>
@@ -8,16 +7,34 @@
 <script lang="ts" setup>
 
 	import {computed} from "vue";
-	import {BulmaAlignments, getBulmaClassesFromProps} from "../../../types";
+	import {BulmaAlignment, toBulmaAlignmentClasses} from "../../../types";
 
 	const props = withDefaults(defineProps<{
-		isGrouped?: false | BulmaAlignments;
+		/**
+		 * The alignment of the control group
+		 */
+		alignment?: BulmaAlignment;
 		/** make form controls [appear] attached together. Not recommended with a label*/
-		hasAddons?: boolean | BulmaAlignments
-		//todo: add is multiline support
+		hasAddons?: boolean
+		/** wrap controls. Not compatible with {@link hasAddons} */
+		isMultiline?: boolean
 	}>(), {});
 
-	const classes = computed(() => getBulmaClassesFromProps(props));
+	const classes = computed(function () {
+		const classes = []
+		if(!props.hasAddons) {
+			classes.push('is-grouped')
+			classes.push('is-grouped-' + toBulmaAlignmentClasses(props.alignment).substring(3))
+			if(props.isMultiline)
+				classes.push('is-grouped-multiline')
+		}
+		else {
+			classes.push('has-addons')
+			classes.push('has-addons-' + toBulmaAlignmentClasses(props.alignment).substring(3))
+		}
+		return classes
+			;
+	});
 
 </script>